requirements.txt
# Core ML + NLP
torch==2.7.1
sentence-transformers==4.1.0
scikit-learn==1.7.0
numpy==1.26.4
pandas==2.2.2

fastapi-0.115.13

# Optional: for fuzzy title matching
rapidfuzz==3.13.0
